Rename files and update tests
